+++ src/components/IngredientFinder.vue
@@ -16,6 +16,9 @@
                             <span>{{ item.name }}</span>
                             <small>{{ item.category }}</small>
                         </div>
+                        <svg v-show="isSelected(item)" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+                            <path d="M12 22C6.47715 22 2 17.5228 2 12C2 6.47715 6.47715 2 12 2C17.5228 2 22 6.47715 22 12C22 17.5228 17.5228 22 12 22ZM11.0026 16L18.0737 8.92893L16.6595 7.51472L11.0026 13.1716L8.17421 10.3431L6.75999 11.7574L11.0026 16Z" fill="currentColor"></path>
+                        </svg>
                     </a>
                     <a v-show="currentQuery" href="#" class="ingredient-finder__options__create" @click.prevent="newIngredient">
                         {{ $t('ingredient.dialog.search-not-found') }} {{ $t('ingredient.dialog.create-ingredient', { name: currentQuery }) }}
@@ -42,6 +45,12 @@ export default {
     },
     props: {
         modelValue: {
+            type: Object,
+            default() {
+                return {}
+            }
+        },
+        cocktailIngredient: { // Temp workaround until cocktail form flow is fixed with v-models
             type: Object,
             default() {
                 return {}
@@ -76,7 +85,7 @@ export default {
             searchClient: instantMeiliSearch(
                 appState.bar.search_driver_host,
                 appState.bar.search_driver_api_key,
-            ),
+            ).searchClient,
         }
     },
     methods: {
@@ -87,6 +96,9 @@ export default {
 
             this.$emit('update:modelValue', ing)
             this.$emit('ingredientSelected', ing)
+        },
+        isSelected(ing) {
+            return this.cocktailIngredient.ingredient_id == ing.id;
         },
         newIngredient() {
             this.isLoading = true
@@ -147,6 +159,7 @@ export default {
     padding: 0.25rem 0.5rem;
     border-radius: var(--radius-1);
     text-decoration: none;
+    align-items: center;
 }
 
 .ingredient-finder__options a:hover {
@@ -155,6 +168,13 @@ export default {
 
 .dark-theme .ingredient-finder__options a:hover {
     background-color: var(--clr-dark-main-700);
+}
+
+.ingredient-finder__options a svg {
+    width: 24px;
+    height: 24px;
+    margin-left: auto;
+    flex-shrink: 0;
 }
 
 .ingredient-finder__options__content {