+++ src/components/SiteAutocomplete.vue
@@ -76,7 +76,7 @@ export default {
         this.searchClient = instantMeiliSearch(
             appState.bar.search_driver_host,
             appState.bar.search_driver_api_key,
-        )
+        ).searchClient
     },
     mounted() {
         this.$nextTick(() => {
@@ -218,7 +218,7 @@ export default {
     font-size: 0.7rem;
     letter-spacing: 1px;
     text-transform: uppercase;
-    font-weight: bold;
+    font-weight: var(--fw-bold);
     margin: 0.75rem 0;
 }
 </style>